+++ src/main/java/org/nrg/containers/model/xnat/Assessor.java
@@ -144,6 +144,11 @@ public class Assessor extends XnatModelObject {
         return new Project(xnatImageassessordataI.getProject(), userI);
     }
 
+    public Session getSession(final UserI userI) {
+        loadXnatImageassessordataI(userI);
+        return new Session(xnatImageassessordataI.getImagesessionId(), userI);
+    }
+
     public void loadXnatImageassessordataI(final UserI userI) {
         if (xnatImageassessordataI == null) {
             xnatImageassessordataI = XnatImageassessordata.getXnatImageassessordatasById(id, userI, false);