+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Not yet released.
 * [CS-458][] `GET /commands/available` includes command and wrapper labels
 * Manually refresh resource catalog when uploading resources. (We used to rely on the Catalog Service to do this, but it doesn't anymore.)
 
+### Bugfixes
+
+* [CS-488][] Allow wrappers to derive `Session` inputs from `Assessor` inputs, in the same way you can derive `Session`s from `Scan`s.
+
 ### Other
 
 * [CS-480][] Deprecate `Container.ContainerMount.inputFiles`. Having a list of input files for a mount is nice during command resolution, but it doesn't make much sense to store that list. As of now no new containers will have anything in their mounts' `inputFiles`. Old containers will still report their values for `inputFiles` for now, but this may change in the future.
@@ -21,6 +25,7 @@ Not yet released.
 [CS-457]: https://issues.xnat.org/browse/CS-457
 [CS-458]: https://issues.xnat.org/browse/CS-458
 [CS-480]: https://issues.xnat.org/browse/CS-480
+[CS-488]: https://issues.xnat.org/browse/CS-488
 
 ## 1.5.1